Move ClubsEnums into enums folder
