+++ src/common/index.ts
@@ -1,5 +1,5 @@
 export * from "./enums/AgeGradeEnums";
-export * from "./ClubsEnums";
+export * from "./enums/ClubsEnums";
 export * from "./refresh";
 export * from "./SearchParams";
 export * from "./SeriesID";
